Did bunch of stuff to find multithreaded garbage collection:
1. Moved e_env to be a local register.  I'm not sure if this is totally the
right thing, but it seems to have stopped most of the crashing.

2. Changed allocate macros in data.h to first aquire alloc lock and then
check to see if enough memory is available.  This prevented a race
condition when two or more threads entered alloc when there was enough
memory for each, but not enough for all of them.

3. Decremented e_pc before it was touched by the garbage collector.  This
is because the GC, in mulithreaded, can be run on branch and funcall
instructions which could leave e_pc just outside of the code block.